requirements.txt
synapse
pytest
python-dotenv